Adding insert view for data of the raspeedi table
