+++ dashboard/forms.py
@@ -1,0 +1,12 @@
+from django.forms.utils import ErrorList
+
+
+class ParaErrorList(ErrorList):
+
+    def __str__(self):
+        return self.as_divs()
+
+    def as_divs(self):
+        if not self:
+            return ''
+        return '<div>%s</div>' % ''.join(['<p class="text-danger">* %s</p>' % e for e in self])
